Added basic searching
A basic search implementation. Search for offices, assignments, skills and users
using a combination of offices, assignments, skills, users, roles, skill levels and skill years.

All searches do an intersection of the results for each tag, so it is required that a result is matched by
every tag in the search.